#0000
Gradle meet willow...

Don't panic! This is NOT replacing maven with gradle at this time. Continue to use maven as you have always. I had some time while watching something on TV, and thought I'd throw this together quickly to see how it worked. So far, this is just all the dependencies translated to gradle-speak and nothing else. First we'll get angel 100% happy and releasing before doing anything to willow.



git-svn-id: https://svn.ish.com.au/ish/willow/trunk@21755 fb2e276d-fae9-0310-b578-b71ac85840b9