#0000 Remove the version number on the Libraries folder which makes deployment in the cloud slightly simpler.
Also, runtime dependencies are not the same thing as compile dependencies.

In the most common case, the artifacts needed at compile time are a subset of those needed at runtime. For example, let's say that a program called app uses library foo, and library foo internally uses library bar. Then only foo is needed to compile app, but both foo and bar are needed to run it. This is why by default, everything that you put on Gradle's compile configuration is also visible on its runtime configuration, but the opposite isn't true.