OD-12977 Created spec for Canvas closure The spec requires:
name - allows for multple canvas instances

enrolment - the enrolment being passed in the canvas instance

create_section - a flag used for Canvas Section creation. If this flag is set to TRUE and there is no Canvas section equivalent to the CourseClass being enrolled in, a new Section is created with the name being the CourseClass uniqueCode. if FALSE and no equivalent Section exists, an exception is thrown

create_student - a flag used for Canvas User creation. If this flag is set to TRUE and there is no Canvas User equivalent to the Student being enrolled, a new User is created. If FALSE and no quivalent User exists, an exception is thrown